refactor/rename all helper functions and update comments
