+++ agents/craftassist/tests/test_memory.py
@@ -5,7 +5,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import unittest
 
 import droidlet.base_util
-import droidlet.lowlevel.minecraft.shape_helpers
+import droidlet.lowlevel.minecraft.shape_util
 import droidlet.lowlevel.minecraft.shapes
 from droidlet.memory.craftassist.mc_memory import MCAgentMemory
 from droidlet.shared_data_struct.craftassist_shared_utils import MOBS_BY_ID
