+++ agents/craftassist/tests/test_with_model.py
@@ -6,7 +6,7 @@ import os
 import unittest
 
 import droidlet.base_util
-import droidlet.lowlevel.minecraft.shape_helpers
+import droidlet.lowlevel.minecraft.shape_util
 import droidlet.lowlevel.minecraft.shapes
 from droidlet.lowlevel.minecraft.mc_util import euclid_dist
 from agents.craftassist.tests.base_craftassist_test_case import BaseCraftassistTestCase
